Also show the MusicBrainz lookup ID when retagging.
This information can be quite helpful when a user does automated ripping
and in a second step retags the images to the correct release (using the
-R option).  Using this approach it sometimes turns out that the
specific needs to be added to MusicBrainz first or the disc id needs to
be added to an existent release.  In this case the user needs the lookup
id.