Remove "misc/pep8.py" (outdated PEP8 style checker)
pep8.py was added in 2012. Since then, we now have flake8 and a number
of other really good PEP 8 (and other style, best practice, etc.) code
checkers/linters. It doesn't make sense to track this directly in the
repository. If we want to do automated code style/sanity checking, we
should add a dependency on an actually maintained program/script rather
than keeping it locally in the repository.