Update failing AccurateRipResponse tests [2/2]
The failures seem to be caused by a remote metadata change (metadata from AccurateRip about a release ID changed) which means some of the asserts related to a specific release ID are failing. I've fixed this issue with two commits.
This is part 2: here I've replaced the outdated AccurateRip bin file with a freshly downloaded one.