Use PyYAML for formatting the rip .log file
The .log file is intended to be in YAML, but until now we have
"manually" formatted this output, which is prone to errors.

This new approach generates a dictionary containing the log information
and then calls to PyYAML's YAML implementation to format it out to a
string which can then be saved to a file.

Note that key order of dictionaries prior to Python 3.7 is
non-deterministic (or not guaranteed to be) and thus may need
further tweaking.
Also, PyYAML presently only supports YAML 1.1. It may be needed/wanted
to switch to ruamel.yaml[1] instead for YAML 1.2 support.

[1] https://yaml.readthedocs.io/

Signed-off-by: Frederik “Freso” S. Olesen <freso.dk@gmail.com>