add configuration option to pass extra args to flac encode
Add an argument to the whipper cd rip comand that can be provided on the
command line or in the whipper.cd.rip section of the config file.

It is retrieved inside the _ripIfNotRipped() private function, split
into a list and passed down through to program/flac.py, which actually
calls the flac binary.

Add debug logging to that function so the actual command shows up in the
debug log.

Add a unittest to ensure that arguments passed to flac.encode() work as
expected.

Tested ripping a disc with --extra-flac-encode-args="--best -e
--no-padding" from both the command line and config file as well as with
the default empty value.

Signed-off-by: Kevin Mitchell <kevmitch@gmail.com>